refactor: typed content stores (e.g. FilesystemContentStore) now inherits from v3 ContentStore interface
