+++ spring-content-azure-storage/src/main/java/org/springframework/content/azure/store/AzureStorageContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.azure.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface AzureStorageContentStore<E,CID extends Serializable> extends ContentStore<E,CID> {
 }
