+++ spring-content-fs/src/main/java/org/springframework/content/fs/store/FilesystemContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.fs.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface FilesystemContentStore<I, CID extends Serializable> extends ContentStore<I, CID> {
 }
