+++ spring-content-gcs/src/main/java/org/springframework/content/gcs/store/GCPStorageContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.gcs.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface GCPStorageContentStore<E,CID extends Serializable> extends ContentStore<E,CID> {
 }
