+++ spring-content-jpa/src/main/java/org/springframework/content/jpa/store/JpaContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.jpa.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface JpaContentStore<T, CID extends Serializable> extends ContentStore<T, CID> {
 }
