+++ spring-content-mongo/src/main/java/org/springframework/content/mongo/store/MongoContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.mongo.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface MongoContentStore<E, CID extends Serializable> extends ContentStore<E, CID> {
 }
