+++ spring-content-rest/src/test/java/internal/org/springframework/content/rest/it/http_405/MethodNotAllowedExceptionIT.java
@@ -25,8 +25,8 @@ import org.springframework.content.commons.annotations.ContentId;
 import org.springframework.content.commons.annotations.ContentLength;
 import org.springframework.content.commons.annotations.MimeType;
 import org.springframework.content.commons.property.PropertyPath;
-import org.springframework.content.commons.repository.SetContentParams;
-import org.springframework.content.commons.repository.UnsetContentParams;
+import org.springframework.content.commons.store.SetContentParams;
+import org.springframework.content.commons.store.UnsetContentParams;
 import org.springframework.content.commons.store.ContentStore;
 import org.springframework.content.fs.config.EnableFilesystemStores;
 import org.springframework.content.fs.io.FileSystemResourceLoader;
