+++ spring-content-s3/src/main/java/org/springframework/content/s3/store/S3ContentStore.java
@@ -2,7 +2,7 @@ package org.springframework.content.s3.store;
 
 import java.io.Serializable;
 
-import org.springframework.content.commons.repository.ContentStore;
+import org.springframework.content.commons.store.ContentStore;
 
 public interface S3ContentStore<E,CID extends Serializable> extends ContentStore<E,CID> {
 }
